+++ .github/workflows/release.yaml
@@ -1,4 +1,4 @@
-name: Yasca Release
+name: Zasca Release
 on:
   push:
     branches: [ main ]
@@ -18,10 +18,10 @@ jobs:
         run: pip install -r requirements_dev.txt
 
       - name: Flake8 styles
-        run: python -m flake8 ./yasca
+        run: python -m flake8 ./zasca
 
       - name: Bandit security scan
-        run: python -m bandit -r ./yasca
+        run: python -m bandit -r ./zasca
 
       - name: Safety dependency scan
         run: python -m safety check
@@ -50,8 +50,8 @@ jobs:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         with:
           upload_url: ${{ steps.create_release.outputs.upload_url }}
-          asset_path: ./dist/yasca-${{ env.VERSION }}-py3-none-any.whl
-          asset_name: Yasca Wheel
+          asset_path: ./dist/zasca-${{ env.VERSION }}-py3-none-any.whl
+          asset_name: Zasca Wheel
           asset_content_type: application/x-python-wheel
 
       - name: Publish
@@ -63,6 +63,6 @@ jobs:
 
       - name: Build and push docker image
         run: |
-          docker build -f Dockerfile-release -t javidr/yasca:latest -t javidr/yasca:${{ env.VERSION }} --build-arg yasca_version=${{ env.VERSION }} . --progress=plain
+          docker build -f Dockerfile-release -t javidr/zasca:latest -t javidr/zasca:${{ env.VERSION }} --build-arg zasca_version=${{ env.VERSION }} . --progress=plain
           echo ${{ secrets.DOCKERHUB_PASSWORD }} | docker login -u ${{ secrets.DOCKERHUB_USERNAME }} --password-stdin
-          docker push javidr/yasca -a
+          docker push javidr/zasca -a
