+++ Dockerfile
@@ -1,5 +1,5 @@
 FROM javidr/yasca_base:latest
-COPY . yasca/
-RUN cd yasca && python setup.py sdist bdist_wheel && pip install dist/*.whl
+COPY . zasca/
+RUN cd zasca && python setup.py sdist bdist_wheel && pip install dist/*.whl
 COPY entrypoint.sh .
 ENTRYPOINT ["/bin/sh","/entrypoint.sh"]
