+++ MANIFEST.in
@@ -1,3 +1,3 @@
-include yasca/data/queries.yaml
-include yasca/template/report_template.html
-include yasca/template/cyclonedx_template.json
+include zasca/data/queries.yaml
+include zasca/template/report_template.html
+include zasca/template/cyclonedx_template.json
