+++ entrypoint.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 set -e
 all_args=$@
-yasca  ${all_args}
+zasca  ${all_args}
