+++ zasca/template/report_template.html
@@ -20,7 +20,7 @@ th {
 
 </head>
 
-<title> SCA report </title>
+<title> Zasca report </title>
 <body>
 <h1>{{ appname.package }}:{{ appname.version }} Report </h1>
 <h2>Time: {{ now }}</h2>
